backend/requirements.txt
django
djangorestframework
djangorestframework-xml
djoser
django-cors-headers
psycopg2
pillow
bleach
djangorestframework-csv
djangorestframework-yaml
django-phone-field
djangorestframework-simplejwt~=5.2.1
psycopg2-binary
joblib
pandas
opencv-python
tensorflow
keras
keras-preprocessing
